Added two Docker Files for Streamlit and Spotify Backstage. I have also re-organized the directory
