+++ .vscode/__init__.py
@@ -1,0 +1,8 @@
+__all__ = ['module1', 'module2']
+
+from .module1 import some_function
+from .module2 import another_function
+
+from package import some_function
+
+
