+++ apt/main.py
@@ -4,9 +4,9 @@ them to MP3.
 """
 
 import sys
-sys.path.append('/path/to/youtube_downloader_directory')
+sys.path.append("C:/ProgramFiles(x86)/YTHelper/YTDownloader")
 
-import youtube_downloader
+import ytdownloader
 try:
     import file_converter
 except ImportError:
@@ -36,17 +36,17 @@ if choice == "1" or choice == "2":
     if choice == "2":
         link = input("Enter the link to the playlist: ")
         print("Downloading playlist...")
-        youtube_downloader.download_playlist(link, quality)
+        ytdownloader.download_playlist(link, quality)
         print("Download finished!")
     if choice == "1":
-        links = youtube_downloader.input_links()
+        links = ytdownloader.input_links()
         for link in links:
-            youtube_downloader.download_video(link, quality)
+            ytdownloader.download_video(link, quality)
 elif choice == "3":
-    links = youtube_downloader.input_links()
+    links = ytdownloader.input_links()
     for link in links:
         print("Downloading...")
-        filename = youtube_downloader.download_video(link, 'low')
+        filename = ytdownloader.download_video(link, 'low')
         print("Converting...")
         file_converter.convert_to_mp3(filename)
 else:
